refactor: User 식별자 Integer로 수정

+++ backend/src/main/java/com/ssafy/aboha/user/domain/User.java
@@ -1,7 +1,12 @@
 package com.ssafy.aboha.user.domain;
 
 import com.ssafy.aboha.common.BaseEntity;
-import jakarta.persistence.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -16,7 +21,7 @@ public class User extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     @Column(name = "nickname", unique = true)
     private String nickname;
