+++ backend/src/main/java/com/ssafy/aboha/user/repository/UserRepository.java
@@ -1,11 +1,10 @@
 package com.ssafy.aboha.user.repository;
 
 import com.ssafy.aboha.user.domain.User;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     boolean existsUserByNickname(String nickname);
